+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - _Created_ `tooltip.lua` component.
 - Context-Menu entries now accept `Resolver Functions` that return dynamically generated values for context-menus during run-time.
 - _Created_ `ContextMenu:Add()` to handle table insertion properly. Can handle multiple entries at once.
+- _Add_ `TargetType` and `restrictTo` options.
 
 ### CHANGED